Dockerized
This is a first step. The model comes from experiment six.

Runs the crawler, but now the config work has to happen.
