+++ cmd/crawler/Makefile
@@ -1,4 +1,5 @@
 # SHELL := /bin/bash
+# FIXME: Move to config
 PORT ?= 8888
 
 # make run
@@ -9,6 +10,19 @@ PORT ?= 8888
 #
 # to change the port of the server. 
 run:	
-	cp ../../vcap.json . 
-	DATABASE_URL=postgresql://postgres@localhost:5432/postgres \
+	# FIXME: Move to config
+	#cp ../../vcap.json . 
+	#DATABASE_URL=postgresql://postgres@localhost:5432/postgres \
 	go run *.go -p ${PORT}
+
+build: 
+	go build -buildvcs=false -o crawler.exe
+
+test:
+	go test
+
+run:
+	go run *.go
+
+check:
+	staticcheck
