+++ cmd/crawler/run-fast.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+export SERVICE=crawler
+
+pushd /home/vcap/app/cmd/${SERVICE}
+    echo Running the $SERVICE
+    make run
+popd
