--- pkg/crawl/const.go
@@ -1,3 +1,0 @@
-package crawl
-
-var const_bucket_s3 = "crawl"
