+++ pkg/crawl/types.go
@@ -1,6 +1,12 @@
 package crawl
 
-import "github.com/riverqueue/river"
+import (
+	"github.com/riverqueue/river"
+	"search.eight/internal/queueing"
+	"search.eight/pkg/procs"
+)
+
+var const_bucket_s3 = "crawl"
 
 type CrawlRequest struct {
 	Scheme string `json:"scheme"`
@@ -23,3 +29,15 @@ func (cr CrawlRequest) InsertOpts() river.InsertOpts {
 }
 
 type CrawlRequestJob = river.Job[CrawlRequest]
+
+type CrawlRequestWorker struct {
+	CacheKeyChannel chan string
+	CacheValChannel chan string
+	CacheInsChannel chan map[string]string
+	CleanHtmlClient *queueing.River
+	StorageClient   procs.Storage
+
+	river.WorkerDefaults[CrawlRequest]
+}
+
+type CrawlWorker = river.Worker[CrawlRequest]
